add utils.py in setings
