+++ src/adminsettings/views.py
@@ -1,5 +1,5 @@
+from .utils import set_settings, set_default, get_settings
 from .mixins import SettingsMixin, RedirectMixin
-from .settings import allsettings
 from django.contrib import messages
 
 
@@ -15,7 +15,7 @@ class SettingsChangeDoneView(SettingsMixin):
     change settings and redirect to change settings
     """
     def get(self, request, *args, **kwargs):
-        allsettings.set_settings(request.POST)
+        set_settings(request.POST)
         messages.success(request, "Settings changed successfully ...")
         return RedirectMixin.as_view(redirect_url='settings:settings_change')(request)
 
@@ -25,7 +25,7 @@ class SettingsResetView(SettingsMixin):
     reset settings and redirect to change settings
     """
     def get(self, request, *args, **kwargs):
-        allsettings.set_default()
-        allsettings.set_settings(allsettings.get_settings())
+        set_default()
+        set_settings(get_settings())
         messages.success(request, "Settings reset successfully ...")
         return RedirectMixin.as_view(redirect_url='settings:settings_change')(request)
